Don't ignore keys just because xkey.state != 0
Any keys pressed while e.g. NumLock is active have non-zero state.
Combinations like Alt+Tab will never be received by this program anyway
because they'll be intercepted by the window manager. So it is only
necessary to ignore the press of the Alt itself, not of the Tab with
non-zero state.